backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
cryptography==43.0.1
pydantic==2.9.2
python-multipart==0.0.9
zeep==4.3.1
lxml==5.3.0
# xmlsec==1.3.13
# Se comenta temporalmente en Windows por errores al construir el wheel (paths largos).
# La autenticación se hará vía SAT_USE_SATCFDI=1 y/o SAT_MANUAL_AUTH=1.
# En entorno Linux/WSL se puede reinstalar esta línea para firma WS-Security con zeep.
pyOpenSSL==24.2.1
beautifulsoup4==4.12.3
packaging==24.1
